organised the codebase
