+++ index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css" integrity="sha384-1e3cf9f44e9f0e8a3e1b9e9b0f60e9b5f7b5b5b5f7b5b5b5f7b5b5b5b5b5b5" crossorigin="anonymous" />
 
-    <title>BootStrap!</title>
+    <title>Connect Hub</title>
   </head>
   <body>
     <div id="root"></div>
