+++ package.json
@@ -13,7 +13,8 @@
     "flowbite": "^2.4.1",
     "flowbite-react": "^0.10.1",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.25.1"
   },
   "devDependencies": {
     "@types/react": "^18.3.3",
