--- src/Card.jsx
@@ -1,32 +1,0 @@
-import React from 'react'
-const obj=[
-    {img:"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGOsnw1Zlur9Rx0Z7-MjRg2b_CMEUQMYYLAA&svite-project\src\TheStrayDog.jpeg",
-    },
-    {img:"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGOsnw1Zlur9Rx0Z7-MjRg2b_CMEUQMYYLAA&svite-project\src\TheStrayDog.jpeg",
-    },
-    {img:"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGOsnw1Zlur9Rx0Z7-MjRg2b_CMEUQMYYLAA&svite-project\src\TheStrayDog.jpeg",
-    },
-    {img:"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGOsnw1Zlur9Rx0Z7-MjRg2b_CMEUQMYYLAA&svite-project\src\TheStrayDog.jpeg",
-    },
-    {img:"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGOsnw1Zlur9Rx0Z7-MjRg2b_CMEUQMYYLAA&svite-project\src\TheStrayDog.jpeg",
-    },
-    {img:"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGOsnw1Zlur9Rx0Z7-MjRg2b_CMEUQMYYLAA&svite-project\src\TheStrayDog.jpeg",
-    },
-    {img:"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGOsnw1Zlur9Rx0Z7-MjRg2b_CMEUQMYYLAA&svite-project\src\TheStrayDog.jpeg",
-    },
-    {img:"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGOsnw1Zlur9Rx0Z7-MjRg2b_CMEUQMYYLAA&svite-project\src\TheStrayDog.jpeg",
-    },
-    {img:"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGOsnw1Zlur9Rx0Z7-MjRg2b_CMEUQMYYLAA&svite-project\src\TheStrayDog.jpeg",
-    },
-    {img:"https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGOsnw1Zlur9Rx0Z7-MjRg2b_CMEUQMYYLAA&svite-project\src\TheStrayDog.jpeg",
-    }
-]
-
-export const Card = () => {
-  return (
-    <>
-    
-
-    </>
-  )
-}
