+++ src/pages/About.jsx
@@ -1,0 +1,5 @@
+function About() {
+
+}
+
+export default About
