+++ src/pages/Blogs.jsx
@@ -1,0 +1,5 @@
+function Blog() {
+
+}
+
+export default Blog
