+++ src/pages/Connect.jsx
@@ -1,0 +1,5 @@
+function Connect() {
+    
+}
+
+export default Connect
