+++ src/pages/Layout.jsx
@@ -1,0 +1,15 @@
+import { Outlet } from "react-router-dom"
+import Nav from "../components/Nav"
+import Footer from "../components/Footer"
+
+function Layout() {
+    return (
+        <>
+        <Nav/>
+        <Outlet/>
+        <Footer/>
+        </>
+    )
+}
+
+export default Layout
